requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchainhub>=0.1.0
python-dotenv>=1.0.0
flask>=2.3.0
tweepy>=4.14.0
pydantic>=2.0.0
requests>=2.31.0
black>=23.0.0
